Adjustment of the center and orientation of a latent configuration.

adjust_x adjusts the center and orientation of a latent configuration in Bayesian (metric) multidimensional scaling (BMDS).

adjust_x(x)

Arguments

x

Numeric matrix containing the latent configuration.

Value

A list with elements:

x

A real matrix containing the adjusted latent configuration.

Sig_x

The variance and covariance matrix of the adjusted latent configuration.

See also

Author

Sergio Venturini sergio.venturini@unicatt.it

Examples

n <- 100
nr <- 20
nc <- floor(n/nr)
x <- matrix(rnorm(1:n), nrow = nr, ncol = nc)
adj_x <- adjust_x(x)
adj_x$x
#>             [,1]       [,2]        [,3]        [,4]        [,5]
#>  [1,] -0.2980819 -0.6001930  0.08093344  1.45865998 -0.14381830
#>  [2,] -1.8412381  0.1357997 -0.04944892 -0.62554636  0.31165601
#>  [3,] -0.1391565 -0.3812349 -2.54291165  0.27752593  0.50034756
#>  [4,]  1.1758076  1.4265607 -0.05833301 -0.51393303 -0.41919726
#>  [5,]  0.4262719  0.8704502  0.71711832  0.09296141 -1.04256975
#>  [6,]  1.8694554 -0.3680454  0.38048677 -0.94410010 -0.23609385
#>  [7,] -0.8957346 -0.2331256  1.03039193  0.66903616 -0.49976734
#>  [8,]  0.3234637 -0.2853436  0.62134249 -0.62544661  0.62533165
#>  [9,]  0.2931550  1.1365329 -0.57994093  0.18383563 -0.32462411
#> [10,]  0.8880995 -0.8543075  0.83284314 -1.63543124  0.44133320
#> [11,]  2.6916430  0.2431477 -0.52942568  0.94752692  0.04316384
#> [12,]  0.1290841 -1.7027470  0.43567277  0.32890701 -0.28892155
#> [13,]  0.6187716 -0.6598527  1.00756627  1.93680650  0.47287624
#> [14,] -0.3567457 -0.8936184 -1.41398432  0.39849626 -0.04022697
#> [15,]  0.3072303 -0.2562930 -0.71050451 -0.38597251 -0.05610656
#> [16,] -2.6531360  0.8234298 -0.09702484 -0.06235960 -0.96329542
#> [17,] -0.4873137  2.5527591  1.06958336  0.44808766  1.12450571
#> [18,] -0.4345703 -2.0986349  0.53005006 -0.83074250 -0.10616638
#> [19,]  0.4652184  1.4133855 -0.69556218 -0.90909383 -0.05391237
#> [20,] -2.0822236 -0.2686695 -0.02885252 -0.20921770  0.65548567
adj_x$Sig_x
#>               [,1]          [,2]          [,3]          [,4]          [,5]
#> [1,]  1.511139e+00  9.941545e-17 -1.041819e-17 -2.973955e-17 -6.132391e-16
#> [2,]  8.162560e-17  1.176819e+00 -4.180605e-16 -7.425877e-17  5.245029e-16
#> [3,]  5.604543e-17 -5.118342e-16  7.762017e-01 -4.292981e-16  1.391662e-16
#> [4,]  1.146218e-17 -1.001839e-16 -4.301818e-16  7.068672e-01 -5.126613e-16
#> [5,] -5.786303e-16  5.304050e-16  1.390163e-16 -4.491274e-16  2.788775e-01