Estimation of a DMBC model.

dmbc(), the main function of the package, estimates a DMBC model for a given set of S dissimilarity matrices.

dmbc(
  data,
  p = 2,
  G = 3,
  control = dmbc_control(),
  prior = NULL,
  cl = NULL,
  post_all = FALSE
)

Arguments

data

An object of class dmbc_data containing the data to analyze.

p

A length-one numeric vector indicating the number of dimensions of the latent space.

G

A length-one numeric vector indicating the number of cluster to partition the S subjects.

control

A list of control parameters that affect the sampling but do not affect the posterior distribution. See dmbc_control() for more details.

prior

A list containing the prior hyperparameters. See dmbc_prior() for more details.

cl

An optional parallel or snow cluster for use if parallel = "snow". If not supplied, a cluster on the local machine is created for the duration of the dmbc() call.

post_all

A length-one logical vector, which if TRUE applies a further post-processing to the simulated chains (in case these are more than one).

Value

A dmbc_fit_list object.

References

Venturini, S., Piccarreta, R. (2021), "A Bayesian Approach for Model-Based Clustering of Several Binary Dissimilarity Matrices: the dmbc Package in R", Journal of Statistical Software, 100, 16, 1--35, <10.18637/jss.v100.i16>.

See also

Author

Sergio Venturini sergio.venturini@unicatt.it

Examples

if (FALSE) {
data(simdiss, package = "dmbc")

G <- 3
p <- 2
prm.prop <- list(z = 1.5, alpha = .75)
burnin <- 20000
nsim <- 10000
seed <- 2301

set.seed(seed)

control <- list(burnin = burnin, nsim = nsim, z.prop = prm.prop[["z"]],
  alpha.prop = prm.prop[["alpha"]], random.start = TRUE, verbose = TRUE,
  nchains = 2, thin = 10, store.burnin = TRUE, threads = 2,
  parallel = "snow")
sim.dmbc <- dmbc(simdiss, p, G, control)

summary(sim.dmbc, include.burnin = FALSE)

library(bayesplot)
library(ggplot2)
color_scheme_set("teal")
plot(sim.dmbc, what = "trace", regex_pars = "eta")

z <- dmbc_get_configuration(sim.dmbc, chain = 1, est = "mean",
  labels = 1:16)
summary(z)
color_scheme_set("mix-pink-blue")
graph <- plot(z, size = 2, size_lbl = 3)
graph + panel_bg(fill = "gray90", color = NA)
}