Model-Based Clustering of Several Dissimilarity Matrices.

The dmbc package implements a Bayesian algorithm for clustering a set of dissimilarity matrices within a model-based framework. In particular, we consider the case where S matrices are available, each describing the dissimilarities among n objects, possibly expressed by S subjects (judges), or measured under different experimental conditions, or with reference to different characteristics of the objects them- selves. Specifically, we focus on binary dissimilarities, taking values 0 or 1 depending on whether or not two objects are deemed as similar, with the goal of analyzing such data using multidimensional scaling (MDS). Differently from the standard MDS algorithms, we are interested in partitioning the dissimilarity matrices into clusters and, simultaneously, to extract a specific MDS configuration for each cluster. The parameter estimates are derived using a hybrid Metropolis-Gibbs Markov Chain Monte Carlo algorithm. We also include a BIC-like criterion for jointly selecting the optimal number of clusters and latent space dimensions.

For efficiency reasons, the core computations in the package are implemented using the C programming language and the RcppArmadillo package.

The dmbc package also supports the simulation of multiple chains through the support of the parallel package.

Plotting functionalities are imported from the nice bayesplot package. Currently, the package includes methods for binary data only. In future releases routines will be added specifically for continuous (i.e. normal), multinomial and count data.

dmbc classes

The dmbc package defines the following new classes:

dmbc_data:

defines the data to use in a DMBC model.

dmbc_model:

defines a DMBC model.

dmbc_fit:

defines the results of a DMBC analysis for a single MCMC chain.

dmbc_fit_list:

defines the results of a DMBC analysis for multiple MCMC chains.

dmbc_ic:

defines the results of the computation of the information criterion for a DMBC analysis.

dmbc_config:

defines the estimate of the latent configuration for a DMBC analysis.

The package includes print, summary and plot methods for each one of these classes.

Resources

Bug reports:

If you have noticed a bug that needs to be fixed, please let us know at the dmbc issue tracker on GitHub:

https://github.com/sergioventurini/dmbc/issues/.

General questions and help:

To ask a question about dmbc send and email to:

sergio.venturini@unicatt.it.

References

Venturini, S., Piccarreta, R. (2021), "A Bayesian Approach for Model-Based Clustering of Several Binary Dissimilarity Matrices: the dmbc Package in R", Journal of Statistical Software, 100, 16, 1–35, <10.18637/jss.v100.i16>.

See also