An S4 class to represent the latent configuration estimate for a DMBC model.

An S4 class to represent the the latent configuration estimate for a DMBC model.

Slots

Z.est

An array containing the estimate of the latent configuration for a DMBC model.

Z.sd

An array containing the standard deviation of the latent configuration for a DMBC model.

cluster

A numeric vector providing the estimated group membership for the S subjects in the data.

est

A length-one character vector providing the estimate type returned in Z.est. Possible values are mean (posterior mean), median (posterior median), ml (maximum likelihood) and map (maximum-a-posteriori).

n

A length-one numeric vector providing the number of objects.

p

A length-one numeric vector providing the number of latent dimensions.

S

A length-one numeric vector providing the number of subjects.

G

A length-one numeric vector providing the number of clusters.

family

An object of class list; named list with elements representing the parameter estimates corresponding to different values of p and G.

chain

A length-one numeric vector representing the ID of the MCMC chain used to compute the estimates.

labels

A character vector for the (optional) strings to use in the plots for labeling the objects.

References

Venturini, S., Piccarreta, R. (2021), "A Bayesian Approach for Model-Based Clustering of Several Binary Dissimilarity Matrices: the dmbc Package in R", Journal of Statistical Software, 100, 16, 1--35, <10.18637/jss.v100.i16>.

Examples

showClass("dmbc_config")
#> Class "dmbc_config" [package "dmbc"]
#> 
#> Slots:
#>                                                                             
#> Name:      Z.est      Z.sd   cluster       est         n         p         S
#> Class:     array     array   numeric character   numeric   numeric   numeric
#>                                               
#> Name:          G    family     chain    labels
#> Class:   numeric character   numeric character