Auxiliary Function for Controlling DMBC Model Fitting

dmbc_control() is an auxiliary function as user interface for dmbc() fitting. Typically only used when calling the dmbc() function. It is used to set parameters that affect the sampling but do not affect the posterior distribution.control_dmbc() is an alias for dmbc_control().check_control() is an auxiliary function that verifies the correctness of the controls provided before a DMBC is fitted with dmbc().

dmbc_control(
  nsim = 5000,
  burnin = 10000,
  thin = 1,
  nchains = 1,
  threads = 1,
  seed = NULL,
  parallel = "no",
  z.prop = 1.5,
  alpha.prop = 0.75,
  random.start = TRUE,
  partition = NULL,
  method = "manhattan",
  procrustes = TRUE,
  relabel = TRUE,
  store.burnin = TRUE,
  verbose = FALSE
)

control_dmbc(
  nsim = 5000,
  burnin = 10000,
  thin = 1,
  nchains = 1,
  threads = 1,
  seed = NULL,
  parallel = "no",
  z.prop = 1.5,
  alpha.prop = 0.75,
  random.start = TRUE,
  partition = NULL,
  method = "manhattan",
  procrustes = TRUE,
  relabel = TRUE,
  store.burnin = TRUE,
  verbose = FALSE
)

check_control(control)

Arguments

nsim

A length-one numeric vector for the number of draws to be taken from the posterior distribution.

burnin

A length-one numeric vector for the number of initial MCMC iterations (usually to be discarded).

thin

A length-one numeric vector for the number of iterations between consecutive draws.

nchains

A length-one numeric vector for the number of parallel chains to run.

threads

A length-one numeric vector for the number of chains to run. If greater than 1, package parallel is used to take advantage of any multiprocessing or distributed computing capabilities that may be available.

seed

An integer scalar. If supplied, provides the random number seed.

parallel

A length-one character vector indicating the type of parallel operation to be used (if any). Possible values are multicore (which works only on Unix/mcOS), snow and no (i.e. serial instead of parallel computing).

z.prop

A length-one numeric vector providing the standard deviation of the proposal distribution for the jump in the individual latent space position.

alpha.prop

A length-one numeric vector providing the standard deviation of the proposal distribution for the jump in the individual random effect value.

random.start

A length-one logical vector. If TRUE the starting values are drawn randomly, otherwise a user-defined starting partition must be provided through the partition argument.

partition

A length-one numeric vector providing the user-defined starting partition.

method

A length-one character vector that specifies the distance measure to use in determining the initial partition. Allowed values are those from the dist() function.

procrustes

A length-one logical vector. If TRUE the simulated MCMC chains are post-processed through a Procrustes transformation.

relabel

A length-one logical vector. If TRUE the simulated MCMC chains are relabelled to address the label-switching problem.

store.burnin

A logical scalar. If TRUE, the samples from the burnin are also stored and returned.

verbose

A logical scalar. If TRUE, causes information to be printed out about the progress of the fitting.

control

A list of control options.

Value

A named list with the control options as components.

See also

Author

Sergio Venturini sergio.venturini@unicatt.it

Examples

if (FALSE) { # \dontrun{
data(simdiss, package = "dmbc")
# Shorter run than default.
sim.fit <- dmbc(simdiss,
  control = dmbc_control(burnin = 1000, nsim = 2000, thin = 5, verbose = TRUE))
} # }