Function to compute the starting values before fitting a DMBC models.

dmbc_init() is the main function that estimates a DMBC model.

dmbc_init(D, p, G, family, random.start, method, partition)

Arguments

D

A list whose elements are the dissimilarity matrices corresponding to the judgments expressed by the S subjects/raters. These matrices must be defined as a dist object.

p

A length-one numeric vector indicating the number of dimensions of the latent space.

G

A length-one numeric vector indicating the number of cluster to partition the S subjects.

family

A length-one character vector representing the type of data to analyze. Currently, it accepts only the 'binomial' value, but future developments will include the possibility to analyze continuous, multinomial and count data.

random.start

A length-one logical vector. If TRUE the starting values are drawn randomly, otherwise.

method

A length-one character vector specifying the distance measure to use in determining the initial partition. Allowed values are those from the dist() function.

partition

A length-one numeric vector providing the user-defined starting partition.

Value

A named list with the following items:

z:

array of latent coordinates starting values

x:

numeric vector of initial cluster memberships

ng:

numeric vector of initial cluster sizes

alpha:

numeric vector of alpha starting values

eta:

numeric vector of eta starting values

sigma2:

numeric vector of sigma2 starting values

lambda:

numeric vector of lambda starting values

References

Venturini, S., Piccarreta, R. (2021), "A Bayesian Approach for Model-Based Clustering of Several Binary Dissimilarity Matrices: the dmbc Package in R", Journal of Statistical Software, 100, 16, 1--35, <10.18637/jss.v100.i16>.

See also

Author

Sergio Venturini sergio.venturini@unicatt.it

Examples

data(simdiss, package = "dmbc")
dmbc_init(simdiss@diss, p = 2, G = 3, family = "binomial", random.start = TRUE)
#> $z
#> , , 1
#> 
#>             [,1]        [,2]
#>  [1,] -0.9609092 -0.06939048
#>  [2,] -0.9053111  0.10188445
#>  [3,] -1.2074960 -0.85160131
#>  [4,]  1.2259221  0.18062189
#>  [5,] -0.9578970 -0.34060105
#>  [6,]  0.4754328  1.90915681
#>  [7,]  0.5972067 -1.09567874
#>  [8,] -1.0581664  0.34012916
#>  [9,]  0.7420469 -2.03452785
#> [10,]  1.3442510 -0.36029923
#> [11,]  0.6527496  2.01891644
#> [12,]  0.9699793 -0.18903301
#> [13,] -0.5803217  0.73545150
#> [14,] -0.5379315  0.23915924
#> [15,]  1.3857844 -0.28054003
#> [16,] -1.1853400 -0.30364779
#> 
#> , , 2
#> 
#>              [,1]        [,2]
#>  [1,] -1.07904836  0.47161451
#>  [2,] -1.33930350  0.26192595
#>  [3,]  0.08422762 -1.60470989
#>  [4,]  1.14117462 -0.08434395
#>  [5,] -1.06129376 -1.69162305
#>  [6,]  1.08570429  0.24952808
#>  [7,] -0.39914498 -0.67819627
#>  [8,] -0.72812229 -1.53159904
#>  [9,] -1.12360778  1.20108593
#> [10,]  1.30362895 -0.27988515
#> [11,]  1.13488581 -0.01606624
#> [12,]  0.11611910  1.50116826
#> [13,]  1.24154230  0.22112901
#> [14,]  0.11803978  1.31020459
#> [15,]  0.70043941 -0.22753641
#> [16,] -1.19524121  0.89730367
#> 
#> , , 3
#> 
#>             [,1]       [,2]
#>  [1,] -1.3456028 -0.8763089
#>  [2,] -1.0485967 -0.8674010
#>  [3,] -1.0076823 -0.1475845
#>  [4,]  1.1606790  0.3862016
#>  [5,] -1.0077162  0.4464319
#>  [6,]  1.0648920 -0.4158996
#>  [7,] -0.5099254  1.2758965
#>  [8,] -0.7334108 -0.3117400
#>  [9,] -0.7833140  2.0319533
#> [10,]  1.1948119  0.9687496
#> [11,]  1.0758043 -0.5669077
#> [12,]  0.8138698  0.6439424
#> [13,]  0.4810764 -1.3770776
#> [14,]  0.5940928 -1.7482888
#> [15,]  1.1056079  0.6659821
#> [16,] -1.0545859 -0.1079494
#> 
#> 
#> $x
#>  [1] 1 1 3 3 2 2 3 1 2 2
#> 
#> $ng
#> 
#> 1 2 3 
#> 3 4 3 
#> 
#> $alpha
#> [1] -1.110953 -1.870867 -1.327167
#> 
#> $eta
#> [1] 1 1 1
#> 
#> $sigma2
#> [1] 0.04295166 0.03882732 0.04043281
#> 
#> $lambda
#> 
#>   1   2   3 
#> 0.3 0.4 0.3 
#>