Provide a graphical summary of a dmbc_fit_list class instance.

Provide a graphical summary of a dmbc_fit_list class instance.

# S4 method for class 'dmbc_fit_list,ANY'
plot(
  x,
  what = "trace",
  pars = character(),
  regex_pars = "lambda",
  include.burnin = FALSE,
  combo = NULL,
  ...
)

Arguments

x

An object of class dmbc_fit_list.

what

A length-one character vector providing the plot type to produce. Admissible values are those provided by the bayesplot package, that is: acf, areas, dens, hex, hist, intervals, neff, pairs, parcoord, recover, rhat, scatter, trace, violin or combo. In particular, combo allows to mix different plot types. For more details see the documentation of the bayesplot package.

pars

An optional character vector of parameter names. If neither pars nor regex_pars is specified, the default is to use all parameters.

regex_pars

An optional regular expression to use for parameter selection. Can be specified instead of pars or in addition to pars.

include.burnin

A length-one logical vector. If TRUE the burnin iterations (if available) are included in the summary.

combo

A character vector providing the plot types to combine (see mcmc_combo).

...

Further arguments to pass on.

Author

Sergio Venturini sergio.venturini@unicatt.it