Adjustment of the center and orientation of a latent configuration.

adjust_x adjusts the center and orientation of a latent configuration in Bayesian (metric) multidimensional scaling (BMDS).

adjust_x(x)

Arguments

x

Numeric matrix containing the latent configuration.

Value

A list with elements:

x

A real matrix containing the adjusted latent configuration.

Sig_x

The variance and covariance matrix of the adjusted latent configuration.

See also

Author

Sergio Venturini sergio.venturini@unicatt.it

Examples

n <- 100
nr <- 20
nc <- floor(n/nr)
x <- matrix(rnorm(1:n), nrow = nr, ncol = nc)
adj_x <- adjust_x(x)
adj_x$x
#>              [,1]       [,2]         [,3]        [,4]        [,5]
#>  [1,]  0.04062548 -0.8546512  1.553545930 -1.19735013  0.03897121
#>  [2,] -0.44825310  1.0274438  0.003494222 -0.83136652  0.63373903
#>  [3,]  1.83891736 -1.4090950  1.322951373  0.13184698  0.54392116
#>  [4,] -0.36353033  0.3877125 -0.381150931 -0.67436913 -0.07076283
#>  [5,] -1.42964752  0.2622913 -2.097173111 -0.55882570 -0.49471040
#>  [6,] -2.77911979 -0.1734218  0.007031628  1.57441487  0.67298386
#>  [7,]  2.52626345 -1.4394794 -1.295284633  1.37730342 -1.17185527
#>  [8,] -0.07776246 -0.2870418 -0.373058003  0.50048446 -0.28568192
#>  [9,]  0.43902089  0.1459469  0.508482821  0.62676591  0.02422896
#> [10,]  1.80096243  1.6454338 -1.285057594  0.27018359  1.63609285
#> [11,] -1.12193596 -0.5630128 -1.189023939 -0.39944833  0.87257784
#> [12,] -0.47929241  0.9532959  0.470159199 -0.25642465 -0.22226329
#> [13,] -0.96465370  1.8720757  1.023907344  0.18945526 -1.48413136
#> [14,]  0.17885013 -1.3147280 -0.753595152 -1.39719671 -0.55271752
#> [15,] -0.11041013  0.5013514  0.705120341  0.15620308 -1.18594780
#> [16,]  0.19478230 -1.3361533  1.281758520 -0.58806586  0.72645565
#> [17,]  2.85297747  1.4198020  0.224024312  0.37477979  0.14607523
#> [18,] -1.81441522 -1.1637920  0.493742744  1.40574855  0.38884903
#> [19,] -0.26431980  1.0884400  0.744955220 -0.09786499  0.19382762
#> [20,] -0.01905909 -0.7624178 -0.964830292 -0.60627389 -0.40965207
adj_x$Sig_x
#>               [,1]          [,2]          [,3]          [,4]          [,5]
#> [1,]  1.865548e+00 -2.525231e-16 -2.470414e-16  5.871736e-16 -3.280304e-16
#> [2,] -2.655953e-16  1.126794e+00  3.927442e-17 -6.463254e-17  6.946301e-17
#> [3,] -2.822282e-16  2.330843e-17  9.802542e-01 -2.790137e-16  9.085649e-17
#> [4,]  6.225751e-16 -6.607414e-17 -2.609040e-16  6.539324e-01  2.126646e-16
#> [5,] -3.934594e-16 -1.398990e-17 -1.565979e-17  1.505948e-16  5.582482e-01